* refactor AudioWrap.js to audiowrap.js
