+++ client/src/components/endRound/PlayerChooser.vue
@@ -33,7 +33,7 @@
 import {createNamespacedHelpers} from 'vuex';
 import Click1Mp3 from '@/assets/audio/click1.mp3';
 import Click2Mp3 from '@/assets/audio/click2.mp3';
-import {AudioWrap} from '@/mixins/AudioWrap';
+import {AudioWrap} from '@/mixins/audiowrap';
 
 const {mapState, mapGetters} = createNamespacedHelpers('room');
