+++ client/src/components/gameShared/ResponseList.vue
@@ -29,7 +29,7 @@
 import {createNamespacedHelpers} from 'vuex';
 import Click1Mp3 from '@/assets/audio/click1.mp3';
 import Click2Mp3 from '@/assets/audio/click2.mp3';
-import {AudioWrap} from '@/mixins/AudioWrap';
+import {AudioWrap} from '@/mixins/audiowrap';
 
 const {mapGetters} = createNamespacedHelpers('game');
