+++ client/src/components/responseMatching/DisputeIcon.vue
@@ -17,7 +17,7 @@ import {createNamespacedHelpers} from 'vuex';
 import NotificationCount from '@/components/gameShared/NotificationCount.vue';
 import ClickMp3 from '@/assets/audio/click2.mp3';
 import {Tooltip} from 'bootstrap';
-import {AudioWrap} from '@/mixins/AudioWrap';
+import {AudioWrap} from '@/mixins/audiowrap';
 
 const {mapGetters, mapState} = createNamespacedHelpers('game');
