+++ client/src/components/responseMatching/MatchCard.vue
@@ -20,7 +20,7 @@
 <script>
 import {createNamespacedHelpers} from 'vuex';
 import ClickMp3 from '@/assets/audio/click2.mp3';
-import {AudioWrap} from '@/mixins/AudioWrap';
+import {AudioWrap} from '@/mixins/audiowrap';
 
 const room = createNamespacedHelpers('room');
 const game = createNamespacedHelpers('game');
