+++ client/src/views/Home.vue
@@ -41,7 +41,7 @@
 <script>
 import {createNamespacedHelpers} from 'vuex';
 import ClickMp3 from '@/assets/audio/click2.mp3'
-import {AudioWrap} from '@/mixins/AudioWrap';
+import {AudioWrap} from '@/mixins/audiowrap';
 
 const click = new AudioWrap(ClickMp3);
